Move all to server js
