--- backend/app.js
@@ -1,19 +1,0 @@
-const express = require("express");
-const app = express();
-const path = require("path");
-
-app.use("/", express.static(path.join(__dirname, "/public")));
-
-app.use("/", require("./routes/root"));
-
-app.all("*", (req, res) => {
-  const filePath = req.accepts("html")
-    ? path.join(__dirname, "views", "404.html")
-    : req.accepts("json")
-    ? { message: "404 Not Found" }
-    : "404 Not Found";
-
-  res.status(404).sendFile(filePath);
-});
-
-module.exports = app;
